separate files and add firebase dependency
