+++ lib/dashboard.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class Dashboard extends StatefulWidget {
+  @override
+  _DashboardState createState() => _DashboardState();
+}
+
+class _DashboardState extends State<Dashboard> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
